+++ amulog/alg/shiso/__init__.py
@@ -1,3 +1,4 @@
 from .shiso import LTGenSHISO
+from .shiso import LTGenSHISO as LTGen
 from .shiso import LTGroupSHISO
 from .shiso import init_ltgen